requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
logyca[oauth_token-api_key_simple_auth]==0.1.18
pydantic==2.8.2
python-dotenv==1.0.1